Hihi

+++ APIs/PTP.Domain/Entities/RouteStation.cs
@@ -1,10 +1,7 @@
-using System.Diagnostics;
-using System.Dynamic;
-
 namespace PTP.Domain.Entities;
 public class RouteStation : BaseEntity
 {
-    public byte Index { get; set; } = 0;
+    public int Index { get; set; } = 0;
     public string Description { get; set; } = string.Empty;
     public double DistanceFromStart { get; set; } = default!;
     public double DistanceToNext { get; set; } = default!;
